Add another Event to catch loot chests during world gen
Another coremod that fires an event when the loottable of a
LockableLootTileEntity is being set allowing us to do replace the chest
after the fact.

This also fixes the minor bug that all chests were being replaced
instead of the ratio given in the config. This was set to 100% for
debugging. Oops.